Fe/feat/booking#63

See merge request s09-final/S09P31C206!274

+++ frontend/app/src/main/java/com/ssafy/booking/ui/booking/bookingSetting/BookingSetDateAndFee.kt
@@ -202,8 +202,8 @@ fun SetDateAndFeeBottomButton(
                         meetingId = App.prefs.getMeetingId()!!,
                         date = date.toString(),
                         fee = bookingViewModel.fee.value!!,
-                        lat = App.prefs.getMeetingLat()!!,
-                        lgt = App.prefs.getMeetingLgt()!!,
+                        lat = App.prefs.getMeetingLat()!!.toDouble(),
+                        lgt = App.prefs.getMeetingLgt()!!.toDouble(),
                         address = App.prefs.getMeetingAddress()!!,
                         location = App.prefs.getMeetingLocation()!!,
                     )
